Break everything up and add faker/mimesis support
