+++ mpan/__init__.py
@@ -1,0 +1,2 @@
+from .helpers import is_valid  # NOQA: F401
+from .mpan import MPAN  # NOQA: F401
