+++ pyproject.toml
@@ -11,7 +11,17 @@ documentation = "https://github.com/limejump/mpan"
 
 
 [tool.poetry.dependencies]
+
 python = "^3.8"
+
+Faker = { version = "^10.0.0", optional = true }
+mimesis = { version = "^5.1.0", optional = true }
+coverage = {extras = ["toml"], version = "^6.2"}
+
+
+[tool.poetry.extras]
+faker = ["Faker"]
+mimesis = ["mimesis"]
 
 
 [tool.poetry.dev-dependencies]
@@ -20,7 +30,7 @@ isort = "^5.7.0"
 flake8 = "^3.8.4"
 ipython = "^7.20.0"
 pre-commit = "^2.10.0"
-pytest = "^6.2.2"
+pytest = "^6.2.5"
 pytest-cov = "^2.11.0"
