+++ tests/test_mpan.py
@@ -1,148 +1,9 @@
-import json
-
-from pathlib import Path
 from unittest import TestCase
 
-from mpan import (
-    MPAN,
-    Distributor,
-    InvalidMPANError,
-    MeterTimeSwitchCode,
-    ProfileClass,
-    is_valid,
-)
+from mpan.exceptions import InvalidMPANError
+from mpan.mpan import MPAN
 
-
-# These can't be handled at all by the library
-UNPARSEABLE = (
-    "Not an MPAN",
-    "42",
-)
-
-# These look legit, but don't pass the validation
-INVALID = (
-    "2499999999990",  # Bad checksum
-    "8699999999991",  # Bad distributor
-    "991112221312345678907",  # Bad profile class
-    "000002221312345678907",  # Bad mtc
-)
-
-
-def get_valid_mpans() -> list:
-    """
-    For privacy reasons, valid MPANs cannot be included in this project
-    directly.  Instead, if you want to test for a list of MPANs you know to be
-    legitimate, you must provide them yourself in a file called
-    `tests/data/valid-mpans.json`.  This file should be a JSON list of strings
-    and/or integers, like this:
-
-      [
-        "1234567890123456789012",
-        "1234567890123",
-        1234567890123456789012,
-        1234567890123
-      ]
-
-    """
-    path = Path(__file__).parent / "data" / "valid-mpans.json"
-    if not path.exists():
-        return []
-    with path.open() as f:
-        return json.load(f)
-
-
-class DistributorTestCase(TestCase):
-    def setUp(self):
-        self.dno = Distributor(identifier="10")
-        self.idno = Distributor(identifier="24")
-
-    def test___str__(self):
-        self.assertEqual(str(self.dno), "10")
-        self.assertEqual(str(self.idno), "24")
-
-    def test___repr__(self):
-        self.assertEqual(repr(self.dno), "Distributor: 10")
-        self.assertEqual(repr(self.idno), "Distributor: 24")
-
-    def test___bool__(self):
-        self.assertTrue(Distributor(identifier="10"))
-        self.assertTrue(Distributor(identifier="24"))
-        self.assertFalse(Distributor(identifier="test"))
-
-    def test_is_valid(self):
-        self.assertTrue(self.dno.is_valid)
-        self.assertTrue(self.idno.is_valid)
-        self.assertFalse(
-            Distributor(identifier="99").is_valid  # The Great One 🇨🇦 🏒
-        )
-        self.assertFalse(Distributor(identifier="test").is_valid)
-
-    def test_area(self):
-        self.assertEqual(self.dno.area, "Eastern England")
-        self.assertIsNone(self.idno.area)
-
-    def test_operator(self):
-        self.assertEqual(self.dno.operator, "UK Power Networks")
-        self.assertIsNone(self.idno.operator)
-
-    def test_participant_id(self):
-        self.assertEqual(self.dno.participant_id, "EELC")
-        self.assertIsNone(self.idno.participant_id)
-
-    def test_gsp_group_id(self):
-        self.assertEqual(self.dno.gsp_group_id, "_A")
-        self.assertIsNone(self.idno.gsp_group_id)
-
-    def test_name(self):
-        self.assertIsNone(self.dno.name)
-        self.assertEqual(self.idno.name, "Envoy")
-
-    def test_licensee(self):
-        self.assertIsNone(self.dno.licensee)
-        self.assertEqual(self.idno.licensee, "Independent Power Networks")
-
-    def test_mpas_operator_id(self):
-        self.assertIsNone(self.dno.mpas_operator_id)
-        self.assertEqual(self.idno.mpas_operator_id, "IPNL")
-
-    def test_is_dno(self):
-        self.assertTrue(self.dno.is_dno)
-        self.assertFalse(self.idno.is_dno)
-
-    def test_is_idno(self):
-        self.assertFalse(self.dno.is_idno)
-        self.assertTrue(self.idno.is_idno)
-
-
-class ProfileClassTestCase(TestCase):
-    def setUp(self):
-        self.pc = ProfileClass(identifier="04")
-
-    def test_is_valid(self):
-        self.assertTrue(self.pc)
-        self.assertFalse(ProfileClass(identifier="test"))
-
-    def test_description(self):
-        self.assertEqual(self.pc.description, "Non-domestic Economy 7")
-        self.assertEqual(ProfileClass("").description, None)
-
-
-class MeterTimeSwitchCodeTestCase(TestCase):
-    def test_is_valid(self):
-        self.assertTrue(MeterTimeSwitchCode("001"))
-        self.assertTrue(MeterTimeSwitchCode("999"))
-        self.assertFalse(MeterTimeSwitchCode("1000"))
-        self.assertFalse(MeterTimeSwitchCode("0"))
-        self.assertFalse(MeterTimeSwitchCode("00"))
-        self.assertFalse(MeterTimeSwitchCode("000"))
-        self.assertFalse(MeterTimeSwitchCode("-1"))
-        self.assertFalse(MeterTimeSwitchCode("test"))
-
-    def test_description(self):
-        self.assertEqual(MeterTimeSwitchCode("404").description, "Reserved")
-        self.assertEqual(MeterTimeSwitchCode("").description, None)
-        self.assertEqual(MeterTimeSwitchCode("1000").description, None)
-        self.assertEqual(MeterTimeSwitchCode("test").description, None)
+from .common import INVALID, UNPARSEABLE, get_valid_mpans
 
 
 class MPANTestCase(TestCase):
@@ -243,18 +104,3 @@ class MPANTestCase(TestCase):
         for string in INVALID:
             with self.subTest(string=string):
                 self.assertRaises(InvalidMPANError, MPAN(string).check)
-
-
-class IsValidTestCase(TestCase):
-    def test_is_valid(self):
-        for string in get_valid_mpans():
-            with self.subTest(string=string):
-                self.assertTrue(is_valid(string), string)
-
-    def test_is_valid_fail(self):
-        for string in INVALID:
-            with self.subTest(string=string):
-                self.assertFalse(is_valid(string))
-        for string in UNPARSEABLE:
-            with self.subTest(string=string):
-                self.assertFalse(is_valid(string))
